refactor: move pkg/{git,tui} to root
