+++ server/middleware.go
@@ -8,9 +8,9 @@ import (
 	"github.com/alecthomas/chroma/lexers"
 	gansi "github.com/charmbracelet/glamour/ansi"
 	"github.com/charmbracelet/lipgloss"
+	"github.com/charmbracelet/soft-serve/git"
 	appCfg "github.com/charmbracelet/soft-serve/internal/config"
-	"github.com/charmbracelet/soft-serve/pkg/git"
-	"github.com/charmbracelet/soft-serve/pkg/tui/common"
+	"github.com/charmbracelet/soft-serve/tui/common"
 	"github.com/charmbracelet/wish"
 	gitwish "github.com/charmbracelet/wish/git"
 	"github.com/gliderlabs/ssh"
