+++ tui/common/error.go
