+++ tui/common/reset.go
