+++ tui/tree/bubble.go
@@ -11,11 +11,11 @@ import (
 	"github.com/charmbracelet/bubbles/viewport"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
+	"github.com/charmbracelet/soft-serve/git"
 	"github.com/charmbracelet/soft-serve/internal/tui/style"
-	"github.com/charmbracelet/soft-serve/pkg/git"
-	"github.com/charmbracelet/soft-serve/pkg/tui/common"
-	"github.com/charmbracelet/soft-serve/pkg/tui/refs"
-	vp "github.com/charmbracelet/soft-serve/pkg/tui/viewport"
+	"github.com/charmbracelet/soft-serve/tui/common"
+	"github.com/charmbracelet/soft-serve/tui/refs"
+	vp "github.com/charmbracelet/soft-serve/tui/viewport"
 	"github.com/dustin/go-humanize"
 )
