+++ tui/viewport/viewport_patch.go
